+++ src/environments/environment.production.ts
@@ -4,7 +4,8 @@ import { Api, EnvironmentModel } from './environment.model';
 // TODO Passer en var d'env
 export const ENVIRONMENT: EnvironmentModel = {
   apisConfiguration: {
-    [Api.Adresse]: { domain: 'https://api-adresse.data.gouv.fr' },
+    [Api.Adresse]: { domain: 'https://api.opencagedata.com' },
+    [Api.AdresseFallback]: { domain: 'https://api-adresse.data.gouv.fr' },
     [Api.Geo]: { domain: 'https://geo.api.gouv.fr' },
     [Api.ConseillerNumerique]: { domain: 'https://api.conseiller-numerique.gouv.fr' }
   },
